renamed allPosts to filteredPosts
