+++ src/containers/posts/FilteredPosts.js
